% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries_years.R
\name{get_countries_years}
\alias{get_countries_years}
\title{get_countries_years}
\usage{
get_countries_years(periods, reporters, token = "",
  dest_folder = getwd(), unzip = FALSE)
}
\arguments{
\item{periods}{a vector of years}

\item{reporters}{a vector of reporter (country) codes}

\item{token}{bulk download API access token}

\item{dest_folder}{folder to save file in format id-year.<suffix>. Defaults to working directory}

\item{unzipwhether}{to unzip the downloaded file.}
}
\description{
Download complete trade data for a set of countries and years
}
