% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reporters}
\alias{reporters}
\title{Comtrade reporter country codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254 rows and 2 columns.}
\source{
Comtrade \url{http://comtrade.un.org}
}
\usage{
reporters
}
\description{
A data frame of 5909 records and 26 variables containing commodity codes and
descriptions. Codes cover the core FFD set and the FFD+ set.
}
\details{
\describe{ \item{id}{country id}
\item{text}{Country name}
}
}
\keyword{datasets}
