% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cncodes}
\alias{cncodes}
\title{Combined Nomenclature commodity codes and FFD categories}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5909 rows and 26 columns.}
\source{
HMRC
}
\usage{
cncodes
}
\description{
A data frame of 5909 records and 26 variables containing commodity codes and
descriptions. Codes cover the core FFD set and the FFD+ set.
}
\details{
\describe{ \item{Comcode}{8 digit commodity code}
\item{Description}{Commodity description} \item{...}{and more to be
documented} }
}
\keyword{datasets}
