% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_year.R
\name{get_country_year}
\alias{get_country_year}
\title{Title}
\usage{
get_country_year(period = "2016", reporter = "152", token = "",
  dest_folder, unzip = FALSE)
}
\arguments{
\item{unzip}{}
}
\description{
Title
}
